quipper: Fix signed/unsigned comparisons
Required by gcc 4.9

TEST="""
sudo USE=next_gcc emerge cross-x86_64-cros-linux-gnu/gcc
cros_workon_make --board=wolf --test platform2
"""
BUG=None

Reviewed-on: https://chromium-review.googlesource.com/222932
Tested-by: David Sharp <dhsharp@chromium.org>
Commit-Queue: Caroline Tice <cmtice@chromium.org>

+++ chromiumos-wide-profiling/perf_reader_test.cc
@@ -291,7 +291,7 @@ TEST(PerfReaderTest, ReadsTraceMetadata) {
   testing::ExamplePerfFileAttr_Tracepoint(73).WriteTo(&input);
 
   // data
-  ASSERT_EQ(input.tellp(), header.data.offset);
+  ASSERT_EQ(static_cast<u64>(input.tellp()), header.data.offset);
   testing::ExamplePerfSampleEvent_Tracepoint().WriteTo(&input);
   ASSERT_EQ(input.tellp(), file_header.data_end());
